Move organization page containers
